# apps/web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IPFS Hyperledger Fabric Verification Backend API",
    "description": "API dokumentasi untuk sistem verifikasi ijazah terdesentralisasi menggunakan Hyperledger Fabric (blockchain) dan IPFS (decentralized storage) dengan database lokal PostgreSQL.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login Issuer Admin",
        "description": "Endpoint untuk otentikasi issuer (admin) dan mendapatkan JWT Token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": { "type": "string", "description": "Username, Email, atau Issuer ID", "example": "admin" },
                  "password": { "type": "string", "example": "admin123" }
                },
                "required": ["identifier", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil, mengembalikan token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": { "type": "string", "example": "eyJhbG..." },
                        "issuer": {
                          "type": "object",
                          "properties": {
                            "issuerId": { "type": "string" },
                            "organizationName": { "type": "string" },
                            "departmentName": { "type": "string" },
                            "mspId": { "type": "string" },
                            "username": { "type": "string" },
                            "email": { "type": "string" },
                            "isActive": { "type": "boolean" },
                            "status": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Registrasi Issuer Admin Baru",
        "description": "Mendaftarkan akun issuer admin baru beserta kredensial login (hanya di database, belum didaftarkan ke blockchain ledger).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "issuerId": { "type": "string", "example": "NEW_ISSUER" },
                  "organizationName": { "type": "string", "example": "New University" },
                  "departmentName": { "type": "string", "example": "Academic Office" },
                  "mspId": { "type": "string", "example": "Org1MSP" },
                  "username": { "type": "string", "example": "newadmin" },
                  "email": { "type": "string", "example": "admin@new.ac.id" },
                  "password": { "type": "string", "example": "securepass" }
                },
                "required": ["issuerId", "organizationName", "departmentName", "mspId", "username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Issuer registered successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}